pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "livekit-agents[cartesia,deepgram,openai,silero,turn-detector]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
]
